To implement this, all patients at a given age will be binned according to their model score (using quantiles). Each bin is assigned a state, and we are computing the probability for traversing from each state to the next model state. Patients with missing score are also included for this model to reflect actual population numbers

To implement this, all patients at a given age will be binned according to their model score (using quantiles). Each bin is assigned a state, and we are computing the probability for traversing from each state to the next model state. Patients with missing score are also included for this model to reflect actual population numbers

Usage

mldpEHR.disease_markov(
  models,
  outcome,
  step,
  qbins = seq(0, 1, by = 0.05),
  required_conditions = "id==id"
)

Arguments

models

list of prediction models (output of mldpEHR.cv_train_stitch_outcome)

outcome

time from oldest model (first) to target outcome

step

time between prediction models

qbins

quantile bin size of prediction score for which the markov model will define a state

required_conditions

any filter to apply to the patients to filter out from model computation, for example limiting the time window

Value

a list of markov models (per age), each is a list with the following members:

model - matrix containing the probability for each quantile(score) bin to reach each of the target_classes provided in the oldest model.

local.model - data.frame containing the probability for each quantile(score) bin to reach each of the quantile(score) bins of the next model by age.

qbins - bins

target - data frame containing the target bin for this age model (to be used as outcome for the younger age model)

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(ggplot2)
# build base predictor
N <- 1000
patients <- purrr::map(0:5, ~ data.frame(
    id = 1:N,
    sex = rep(1, N),
    age = 80 - .x * 5,
    death = c(rep(NA, 0.2 * N), rep(82, 0.8 * N)),
    disease = rep(rep(c(NA, 81), each = N / 4), 2),
    followup = .x * 5 + 5
)) %>%
    setNames(seq(80, by = -5, length.out = 6))
features <- purrr::map(0:5, ~ data.frame(
    id = 1:N,
    a = c(rnorm(0.2 * N), rnorm(0.8 * N, mean = 2, sd = 1)),
    b = rep(c(rnorm(N / 4), rnorm(N / 4, mean = 3)), 2)
)) %>% setNames(seq(80, by = -5, length.out = 6))
predictors <- mldpEHR.disease_multi_age_predictors(patients, features, 5, 3)
#> 
#>   Training [-----------------------------] 0/3 (  0%) in  0s
#> 
#>   Training [=========>-------------------] 1/3 ( 33%) in  0s
#> 
#>   Training [==================>----------] 2/3 ( 67%) in  1s
#> 
#>   Training [=============================] 3/3 (100%) in  1s
#> 
#> 
#>   Training [-----------------------------] 0/3 (  0%) in  0s
#> 
#>   Training [=========>-------------------] 1/3 ( 33%) in  0s
#> 
#>   Training [==================>----------] 2/3 ( 67%) in  1s
#> 
#>   Training [=============================] 3/3 (100%) in  1s
#> 
#> 
#>   Training [-----------------------------] 0/3 (  0%) in  0s
#> 
#>   Training [=========>-------------------] 1/3 ( 33%) in  0s
#> 
#>   Training [==================>----------] 2/3 ( 67%) in  1s
#> 
#>   Training [=============================] 3/3 (100%) in  1s
#> 
#> 
#>   Training [-----------------------------] 0/3 (  0%) in  0s
#> 
#>   Training [=========>-------------------] 1/3 ( 33%) in  0s
#> 
#>   Training [==================>----------] 2/3 ( 67%) in  1s
#> 
#>   Training [=============================] 3/3 (100%) in  1s
#> 
#> 
#>   Training [-----------------------------] 0/3 (  0%) in  0s
#> 
#>   Training [=========>-------------------] 1/3 ( 33%) in  0s
#> 
#>   Training [==================>----------] 2/3 ( 67%) in  1s
#> 
#>   Training [=============================] 3/3 (100%) in  1s
#> 
#> 
#>   Training [-----------------------------] 0/3 (  0%) in  0s
#> 
#>   Training [=========>-------------------] 1/3 ( 33%) in  0s
#> 
#>   Training [==================>----------] 2/3 ( 67%) in  1s
#> 
#>   Training [=============================] 3/3 (100%) in  1s
#> 
qbins <- seq(0, 1, by = 0.1)
markov <- mldpEHR.disease_markov(predictors, 5, 5, qbins = qbins)
#> Warning: Insufficient stats for age:80 for sbin:no_score for sex:male ,using pop
#> Error in pivot_wider(., id_cols = c(sbin, sex), names_from = outcome,     values_from = est): could not find function "pivot_wider"
prob <- purrr::map2_df(markov, names(markov), ~
    as_tibble(.x$model[[1]], rownames = "sbin") %>%
        mutate(model = .y)) %>%
    mutate(sbin = factor(sbin,
        levels = c("death", 1:length(qbins), "disease", "disease_death", "no_score")
    ))
#> Error in map2(.x, .y, .f, ...): object 'markov' not found
ggplot(prob, aes(x = sbin, y = disease + disease_death)) +
    geom_point() +
    facet_wrap(~model, nrow = 1) +
    theme_bw()
#> Error in ggplot(prob, aes(x = sbin, y = disease + disease_death)): object 'prob' not found